requirements.txt
psycopg2
whitenoise
# django-bootstrap5
# django-bootstrap-icons
# django-htmx
django-session-timeout